cleans up class structure and adds docu
